// apps/web/src/api/generated/models/KnowledgeSearchRow.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * One search hit projected from the bundle index entry.
 */
export type KnowledgeSearchRow = {
    path: string;
    snippet?: (string | null);
    tags?: Array<string>;
    title: string;
    type: string;
};